+++ labelbox/data/annotation_types/__init__.py
@@ -6,9 +6,15 @@ from .geometry import Rectangle
 from .geometry import Geometry
 
 from .annotation import ClassificationAnnotation
-from .annotation import VideoClassificationAnnotation
 from .annotation import ObjectAnnotation
-from .annotation import VideoObjectAnnotation
+from .video import VideoClassificationAnnotation
+from .video import VideoObjectAnnotation
+from .video import DICOMObjectAnnotation
+from .video import GroupKey
+from .video import MaskFrame
+from .video import MaskInstance
+from .video import VideoMaskAnnotation
+from .video import DICOMMaskAnnotation
 
 from .ner import ConversationEntity
 from .ner import DocumentEntity