+++ labelbox/data/annotation_types/annotation.py
@@ -1,7 +1,7 @@
 import abc
 from typing import Any, Dict, List, Optional, Union
 
-from labelbox.data.mixins import ConfidenceNotSupportedMixin, ConfidenceMixin
+from labelbox.data.mixins import ConfidenceMixin
 
 from .classification import Checklist, Dropdown, Radio, Text
 from .feature import FeatureSchema
@@ -15,7 +15,7 @@ class BaseAnnotation(FeatureSchema, abc.ABC):
     extra: Dict[str, Any] = {}
 
 
-class ClassificationAnnotation(BaseAnnotation):
+class ClassificationAnnotation(BaseAnnotation, ConfidenceMixin):
     """Classification annotations (non localized)
 
     >>> ClassificationAnnotation(
@@ -27,10 +27,12 @@ class ClassificationAnnotation(BaseAnnotation):
         name (Optional[str])
         feature_schema_id (Optional[Cuid])
         value (Union[Text, Checklist, Radio, Dropdown])
+        message_id (Optional[str]) Message id for conversational text
         extra (Dict[str, Any])
      """
 
     value: Union[Text, Checklist, Radio, Dropdown]
+    message_id: Optional[str] = None
 
 
 class ObjectAnnotation(BaseAnnotation, ConfidenceMixin):
@@ -53,47 +55,3 @@ class ObjectAnnotation(BaseAnnotation, ConfidenceMixin):
     """
     value: Union[TextEntity, ConversationEntity, DocumentEntity, Geometry]
     classifications: List[ClassificationAnnotation] = []
-
-
-class VideoObjectAnnotation(ObjectAnnotation, ConfidenceNotSupportedMixin):
-    """Video object annotation
-
-    >>> VideoObjectAnnotation(
-    >>>     keyframe=True,
-    >>>     frame=10,
-    >>>     value=Rectangle(
-    >>>        start=Point(x=0, y=0),
-    >>>        end=Point(x=1, y=1)
-    >>>     ),
-    >>>     feature_schema_id="my-feature-schema-id"
-    >>>)
-
-    Args:
-        name (Optional[str])
-        feature_schema_id (Optional[Cuid])
-        value (Geometry)
-        frame (Int): The frame index that this annotation corresponds to
-        keyframe (bool): Whether or not this annotation was a human generated or interpolated annotation
-        segment_id (Optional[Int]): Index of video segment this annotation belongs to
-        classifications (List[ClassificationAnnotation]) = []
-        extra (Dict[str, Any])
-    """
-
-    frame: int
-    keyframe: bool
-    segment_index: Optional[int] = None
-
-
-class VideoClassificationAnnotation(ClassificationAnnotation):
-    """Video classification
-
-    Args:
-        name (Optional[str])
-        feature_schema_id (Optional[Cuid])
-        value (Union[Text, Checklist, Radio, Dropdown])
-        frame (int): The frame index that this annotation corresponds to
-        segment_id (Optional[Int]): Index of video segment this annotation belongs to
-        extra (Dict[str, Any])
-    """
-    frame: int
-    segment_index: Optional[int] = None