+++ labelbox/data/annotation_types/label.py
@@ -7,14 +7,18 @@ from pydantic import BaseModel, validator
 import labelbox
 from labelbox.data.annotation_types.data.tiled_image import TiledImageData
 from labelbox.schema import ontology
-from .annotation import (ClassificationAnnotation, ObjectAnnotation,
-                         VideoClassificationAnnotation, VideoObjectAnnotation)
+from .annotation import (ClassificationAnnotation, ObjectAnnotation)
 from .classification import ClassificationAnswer
-from .data import VideoData, TextData, ImageData
+from .data import AudioData, ConversationData, DicomData, DocumentData, HTMLData, ImageData, MaskData, TextData, VideoData
 from .geometry import Mask
 from .metrics import ScalarMetric, ConfusionMatrixMetric
 from .types import Cuid
+from .video import VideoClassificationAnnotation
+from .video import VideoObjectAnnotation, VideoMaskAnnotation
 from ..ontology import get_feature_schema_lookup
+
+DataType = Union[VideoData, ImageData, TextData, TiledImageData, AudioData,
+                 ConversationData, DicomData, DocumentData, HTMLData]
 
 
 class Label(BaseModel):
@@ -37,10 +41,9 @@ class Label(BaseModel):
         extra: additional context
     """
     uid: Optional[Cuid] = None
-    data: Union[VideoData, ImageData, TextData, TiledImageData]
+    data: DataType
     annotations: List[Union[ClassificationAnnotation, ObjectAnnotation,
-                            VideoObjectAnnotation,
-                            VideoClassificationAnnotation, ScalarMetric,
+                            VideoMaskAnnotation, ScalarMetric,
                             ConfusionMatrixMetric]] = []
     extra: Dict[str, Any] = {}