+++ labelbox/data/serialization/ndjson/objects.py
@@ -2,7 +2,9 @@ from ast import Bytes
 from io import BytesIO
 from typing import Any, Dict, List, Tuple, Union, Optional
 import base64
+
 from labelbox.data.annotation_types.ner.conversation_entity import ConversationEntity
+from labelbox.data.annotation_types.video import VideoObjectAnnotation, DICOMObjectAnnotation
 from labelbox.data.mixins import ConfidenceMixin
 import numpy as np
 
@@ -16,7 +18,8 @@ from ...annotation_types.data import ImageData, TextData, MaskData
 from ...annotation_types.ner import DocumentEntity, DocumentTextSelection, TextEntity
 from ...annotation_types.types import Cuid
 from ...annotation_types.geometry import Rectangle, Polygon, Line, Point, Mask
-from ...annotation_types.annotation import ClassificationAnnotation, ObjectAnnotation, VideoObjectAnnotation
+from ...annotation_types.annotation import ClassificationAnnotation, ObjectAnnotation
+from ...annotation_types.video import VideoMaskAnnotation, DICOMMaskAnnotation, MaskFrame, MaskInstance
 from .classification import NDSubclassification, NDSubclassificationType
 from .base import DataRow, NDAnnotation
 
@@ -28,6 +31,10 @@ class NDBaseObject(NDAnnotation):
 class VideoSupported(BaseModel):
     # support for video for objects are per-frame basis
     frame: int
+
+
+class DicomSupported(BaseModel):
+    group_key: str
 
 
 class _Point(BaseModel):
@@ -136,6 +143,20 @@ class NDFrameLine(VideoSupported):
                    } for pt in line.points])
 
 
+class NDDicomLine(NDFrameLine):
+
+    def to_common(self, name: str, feature_schema_id: Cuid, segment_index: int,
+                  group_key: str) -> DICOMObjectAnnotation:
+        return DICOMObjectAnnotation(
+            frame=self.frame,
+            segment_index=segment_index,
+            keyframe=True,
+            name=name,
+            feature_schema_id=feature_schema_id,
+            value=Line(points=[Point(x=pt.x, y=pt.y) for pt in self.line]),
+            group_key=group_key)
+
+
 class NDPolygon(NDBaseObject, ConfidenceMixin):
     polygon: List[_Point]
 
@@ -180,10 +201,10 @@ class NDRectangle(NDBaseObject, ConfidenceMixin):
                     extra: Dict[str, Any],
                     data: Union[ImageData, TextData],
                     confidence: Optional[float] = None) -> "NDRectangle":
-        return cls(bbox=Bbox(top=rectangle.start.y,
-                             left=rectangle.start.x,
-                             height=rectangle.end.y - rectangle.start.y,
-                             width=rectangle.end.x - rectangle.start.x),
+        return cls(bbox=Bbox(top=min(rectangle.start.y, rectangle.end.y),
+                             left=min(rectangle.start.x, rectangle.end.x),
+                             height=abs(rectangle.end.y - rectangle.start.y),
+                             width=abs(rectangle.end.x - rectangle.start.x)),
                    data_row=DataRow(id=data.uid, global_key=data.global_key),
                    name=name,
                    schema_id=feature_schema_id,
@@ -212,10 +233,10 @@ class NDFrameRectangle(VideoSupported):
     @classmethod
     def from_common(cls, frame: int, rectangle: Rectangle):
         return cls(frame=frame,
-                   bbox=Bbox(top=rectangle.start.y,
-                             left=rectangle.start.x,
-                             height=rectangle.end.y - rectangle.start.y,
-                             width=rectangle.end.x - rectangle.start.x))
+                   bbox=Bbox(top=min(rectangle.start.y, rectangle.end.y),
+                             left=min(rectangle.start.x, rectangle.end.x),
+                             height=abs(rectangle.end.y - rectangle.start.y),
+                             width=abs(rectangle.end.x - rectangle.start.x)))
 
 
 class NDSegment(BaseModel):
@@ -259,6 +280,31 @@ class NDSegment(BaseModel):
         ])
 
 
+class NDDicomSegment(NDSegment):
+    keyframes: List[NDDicomLine]
+
+    @staticmethod
+    def lookup_segment_object_type(segment: List) -> "NDDicomObjectType":
+        """Used for determining which object type the annotation contains
+        returns the object type"""
+        segment_class = type(segment[0].value)
+        if segment_class == Line:
+            return NDDicomLine
+        else:
+            raise ValueError('DICOM segments only support Line objects')
+
+    def to_common(self, name: str, feature_schema_id: Cuid, uuid: str,
+                  segment_index: int, group_key: str):
+        return [
+            self.segment_with_uuid(
+                keyframe.to_common(name=name,
+                                   feature_schema_id=feature_schema_id,
+                                   segment_index=segment_index,
+                                   group_key=group_key), uuid)
+            for keyframe in self.keyframes
+        ]
+
+
 class NDSegments(NDBaseObject):
     segments: List[NDSegment]
 
@@ -285,6 +331,36 @@ class NDSegments(NDBaseObject):
                    name=name,
                    schema_id=feature_schema_id,
                    uuid=extra.get('uuid'))
+
+
+class NDDicomSegments(NDBaseObject, DicomSupported):
+    segments: List[NDDicomSegment]
+
+    def to_common(self, name: str, feature_schema_id: Cuid):
+        result = []
+        for idx, segment in enumerate(self.segments):
+            result.extend(
+                NDDicomSegment.to_common(segment,
+                                         name=name,
+                                         feature_schema_id=feature_schema_id,
+                                         segment_index=idx,
+                                         uuid=self.uuid,
+                                         group_key=self.group_key))
+        return result
+
+    @classmethod
+    def from_common(cls, segments: List[DICOMObjectAnnotation], data: VideoData,
+                    name: str, feature_schema_id: Cuid, extra: Dict[str, Any],
+                    group_key: str) -> "NDDicomSegments":
+
+        segments = [NDDicomSegment.from_common(segment) for segment in segments]
+
+        return cls(segments=segments,
+                   dataRow=DataRow(id=data.uid, global_key=data.global_key),
+                   name=name,
+                   schema_id=feature_schema_id,
+                   uuid=extra.get('uuid'),
+                   group_key=group_key)
 
 
 class _URIMask(BaseModel):
@@ -341,6 +417,56 @@ class NDMask(NDBaseObject, ConfidenceMixin):
                    confidence=confidence)
 
 
+class NDVideoMasksFramesInstances(BaseModel):
+    frames: List[MaskFrame]
+    instances: List[MaskInstance]
+
+
+class NDVideoMasks(ConfidenceMixin, NDAnnotation):
+    masks: NDVideoMasksFramesInstances
+
+    def to_common(self) -> VideoMaskAnnotation:
+        return VideoMaskAnnotation(
+            frames=self.masks.frames,
+            instances=self.masks.instances,
+            name=self.name,
+            feature_schema_id=self.schema_id,
+        )
+
+    @classmethod
+    def from_common(cls, annotation, data):
+        return cls(
+            data_row=DataRow(id=data.uid, global_key=data.global_key),
+            masks=NDVideoMasksFramesInstances(frames=annotation.frames,
+                                              instances=annotation.instances),
+            name=annotation.name,
+            schema_id=annotation.feature_schema_id,
+        )
+
+
+class NDDicomMasks(NDVideoMasks, DicomSupported):
+
+    def to_common(self) -> DICOMMaskAnnotation:
+        return DICOMMaskAnnotation(
+            frames=self.masks.frames,
+            instances=self.masks.instances,
+            name=self.name,
+            feature_schema_id=self.schema_id,
+            group_key=self.group_key,
+        )
+
+    @classmethod
+    def from_common(cls, annotation, data):
+        return cls(
+            data_row=DataRow(id=data.uid, global_key=data.global_key),
+            masks=NDVideoMasksFramesInstances(frames=annotation.frames,
+                                              instances=annotation.instances),
+            name=annotation.name,
+            schema_id=annotation.feature_schema_id,
+            group_key=annotation.group_key.value,
+        )
+
+
 class Location(BaseModel):
     start: int
     end: int
@@ -392,7 +518,7 @@ class NDDocumentEntity(NDBaseObject, ConfidenceMixin):
                     confidence: Optional[float] = None) -> "NDDocumentEntity":
 
         return cls(text_selections=document_entity.text_selections,
-                   dataRow=DataRow(id=data.uid),
+                   dataRow=DataRow(id=data.uid, global_key=data.global_key),
                    name=name,
                    schema_id=feature_schema_id,
                    uuid=extra.get('uuid'),
@@ -421,7 +547,7 @@ class NDConversationEntity(NDTextEntity):
         return cls(location=Location(start=conversation_entity.start,
                                      end=conversation_entity.end),
                    message_id=conversation_entity.message_id,
-                   dataRow=DataRow(id=data.uid),
+                   dataRow=DataRow(id=data.uid, global_key=data.global_key),
                    name=name,
                    schema_id=feature_schema_id,
                    uuid=extra.get('uuid'),
@@ -453,20 +579,30 @@ class NDObject:
 
     @classmethod
     def from_common(
-        cls, annotation: Union[ObjectAnnotation,
-                               List[List[VideoObjectAnnotation]]],
-        data: Union[ImageData, TextData]
+        cls,
+        annotation: Union[ObjectAnnotation, List[List[VideoObjectAnnotation]],
+                          VideoMaskAnnotation], data: Union[ImageData, TextData]
     ) -> Union[NDLine, NDPoint, NDPolygon, NDRectangle, NDMask, NDTextEntity]:
         obj = cls.lookup_object(annotation)
 
         # if it is video segments
-        if (obj == NDSegments):
-            return obj.from_common(
-                annotation,
-                data,
-                name=annotation[0][0].name,
-                feature_schema_id=annotation[0][0].feature_schema_id,
-                extra=annotation[0][0].extra)
+        if (obj == NDSegments or obj == NDDicomSegments):
+
+            first_video_annotation = annotation[0][0]
+            args = dict(
+                segments=annotation,
+                data=data,
+                name=first_video_annotation.name,
+                feature_schema_id=first_video_annotation.feature_schema_id,
+                extra=first_video_annotation.extra)
+
+            if isinstance(first_video_annotation, DICOMObjectAnnotation):
+                group_key = first_video_annotation.group_key.value
+                args.update(dict(group_key=group_key))
+
+            return obj.from_common(**args)
+        elif (obj == NDVideoMasks or obj == NDDicomMasks):
+            return obj.from_common(annotation, data)
 
         subclasses = [
             NDSubclassification.from_common(annot)
@@ -482,8 +618,21 @@ class NDObject:
     @staticmethod
     def lookup_object(
             annotation: Union[ObjectAnnotation, List]) -> "NDObjectType":
-        if isinstance(annotation, list):
-            result = NDSegments
+
+        if isinstance(annotation, DICOMMaskAnnotation):
+            result = NDDicomMasks
+        elif isinstance(annotation, VideoMaskAnnotation):
+            result = NDVideoMasks
+        elif isinstance(annotation, list):
+            try:
+                first_annotation = annotation[0][0]
+            except IndexError:
+                raise ValueError("Annotation list cannot be empty")
+
+            if isinstance(first_annotation, DICOMObjectAnnotation):
+                result = NDDicomSegments
+            else:
+                result = NDSegments
         else:
             result = {
                 Line: NDLine,
@@ -510,3 +659,4 @@ NDObjectType = Union[NDLine, NDPolygon, NDPoint, NDRectangle, NDMask,
                      NDEntityType, NDDocumentEntity]
 
 NDFrameObjectType = NDFrameRectangle, NDFramePoint, NDFrameLine
+NDDicomObjectType = NDDicomLine