+++ labelbox/utils.py
@@ -1,12 +1,17 @@
 import re
+from urllib.parse import urlparse
 from pydantic import BaseModel
+
+UPPERCASE_COMPONENTS = ['uri', 'rgb']
 
 
 def _convert(s, sep, title):
     components = re.findall(r"[A-Z][a-z0-9]*|[a-z][a-z0-9]*", s)
     components = list(map(str.lower, filter(None, components)))
     for i in range(len(components)):
-        if title(i):
+        if components[i] in UPPERCASE_COMPONENTS:
+            components[i] = components[i].upper()
+        elif title(i):
             components[i] = components[i][0].upper() + components[i][1:]
     return sep.join(components)
 
@@ -30,8 +35,39 @@ def is_exactly_one_set(x, y):
     return not (bool(x) == bool(y))
 
 
+def is_valid_uri(uri):
+    try:
+        result = urlparse(uri)
+        return all([result.scheme, result.netloc])
+    except:
+        return False
+
+
 class _CamelCaseMixin(BaseModel):
 
     class Config:
         allow_population_by_field_name = True
         alias_generator = camel_case
+
+
+class _NoCoercionMixin:
+    """
+    When using Unions in type annotations, pydantic will try to coerce the type
+    of the object to the type of the first Union member. Which results in
+    uninteded behavior.
+
+    This mixin uses a class_name discriminator field to prevent pydantic from
+    corecing the type of the object. Add a class_name field to the class you 
+    want to discrimniate and use this mixin class to remove the discriminator
+    when serializing the object.
+
+    Example:
+        class ConversationData(BaseData, _NoCoercionMixin):
+            class_name: Literal["ConversationData"] = "ConversationData"
+
+    """
+
+    def dict(self, *args, **kwargs):
+        res = super().dict(*args, **kwargs)
+        res.pop('class_name')
+        return res