+++ tests/data/serialization/ndjson/test_text_entity.py
@@ -16,15 +16,3 @@ def test_text_entity_import(filename: str):
     res = list(NDJsonConverter.deserialize(data))
     res = list(NDJsonConverter.serialize(res))
     assert res == data
-
-
-@pytest.mark.parametrize("filename", [
-    "tests/data/assets/ndjson/conversation_entity_import.json",
-    "tests/data/assets/ndjson/conversation_entity_without_confidence_import.json"
-])
-def test_conversation_entity_import(filename: str):
-    with open(filename, 'r') as file:
-        data = json.load(file)
-    res = list(NDJsonConverter.deserialize(data))
-    res = list(NDJsonConverter.serialize(res))
-    assert res == data